+++ migrations/20250507134637_updating_user_table.sql
@@ -1,7 +1,7 @@
 -- +goose Up
 -- +goose StatementBegin
-ALTER TABLE users ADD COLUMN current DOUBLE PRECISION NOT NULL DEFAULT 0;
-ALTER TABLE users ADD COLUMN withdrawn DOUBLE PRECISION NOT NULL DEFAULT 0;
+ALTER TABLE users ADD COLUMN current BIGINT NOT NULL DEFAULT 0;
+ALTER TABLE users ADD COLUMN withdrawn BIGINT NOT NULL DEFAULT 0;
 COMMENT ON COLUMN users.current IS 'Current balance of the user';
 COMMENT ON COLUMN users.withdrawn IS 'Total amount withdrawn by the user';
 -- +goose StatementEnd
